Move enum.utils to models Rename to ClockPosition.
